cleaned up code structure and implemented local storage functions into saved shops page
