+++ client/src/components/ShopList.js
@@ -1,45 +1,11 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import { useLocalStorage } from "../hooks/useLocalStorage";
+import { useLocalStorage, useStorageFunctions } from "../hooks/useLocalStorage";
 import { Palette } from './Palette';
 
 const ShopList = ({ shops, title }) => {
-  // Reminder add the ! back when adding shops is in
-  console.log(shops);
-  const [savedShops, setSavedShops] = useLocalStorage("saved-shops", []);
-
-  const isLiked = (id) => {
-    if(savedShops.filter(item => item._id === id).length > 0){
-      return true
-    }
-    return false
-  }
-
-  
-
-  const likeShop = (e) => {
-    console.log(e.target.id);
-    const {id} = e.target;
-    const likedShop = shops.find(item => item._id === id);
-
-    setSavedShops(prev => {
-      return [...prev, {
-        _id: likedShop._id,
-        businessName: likedShop.businessName,
-        image: likedShop.image,
-      }]
-    })
-  }
-  const unLikeShop = (e) => {
-    console.log("this is being called");
-    const {id} = e.target;
-    setSavedShops(savedShops.filter(item => {
-      return item._id !== id
-    }))
-  }
-
-
+  const {isLiked, likeShop, unLikeShop, } = useStorageFunctions();
 
   if (!shops) {
     return <Container><h2>No Shops Yet 🥲</h2></Container>;
@@ -53,7 +19,7 @@ const ShopList = ({ shops, title }) => {
             <Wrap key={index}>
               <img src={shop.image} alt={shop.businessName} />
                 <div>
-                  <button id={shop._id} onClick={!isLiked(shop._id) ? likeShop : unLikeShop} className={`${isLiked(shop._id) ? "liked" : ""}`}>
+                  <button id={shop._id} onClick={!isLiked(shop._id) ? (() =>likeShop(shop._id)) : (() => unLikeShop(shop._id))} className={`${isLiked(shop._id) ? "liked" : ""}`}>
                     <span></span>
                   </button>
                   <Link to={`/shop/${shop._id}`}>
