+++ client/src/hooks/useLocalStorage.js
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import { useShopContext } from "../contexts/ShopContext";
 
 
 export const useLocalStorage = (key, initialValue) => {
@@ -25,4 +26,44 @@ export const useLocalStorage = (key, initialValue) => {
     }, [key, value])
 
     return [value, setValue]
+}
+
+export const useStorageFunctions = () => {
+    const [savedShops, setSavedShops] = useLocalStorage("saved-shops", []);
+    const { shops } = useShopContext();
+
+    const isLiked = (id) => {
+        if(savedShops.filter(item => item._id === id).length > 0){
+          return true
+        }
+        return false
+    }
+    
+    
+    const likeShop = (id) => {
+        // console.log(e.target.id);
+        // const {id} = e.target;
+        const likedShop = shops.find(item => item._id === id);
+
+        setSavedShops(prev => {
+            return [...prev, {
+                _id: likedShop._id,
+                businessName: likedShop.businessName,
+                image: likedShop.image,
+                liked: true
+
+            }]
+        })
+    }
+
+    const unLikeShop = (id) => {
+        // console.log("this is being called");
+        // const {id} = e.target;
+        setSavedShops(savedShops.filter(item => {
+            return item._id !== id
+        }))
+    }
+
+    return {isLiked, likeShop, unLikeShop, savedShops};
+
 }
